+++ src/components/auth/Login.tsx
@@ -1,12 +1,8 @@
-import {
-  FacebookAuthProvider,
-  GoogleAuthProvider,
-  signInWithEmailAndPassword,
-  signInWithPopup,
-} from "firebase/auth";
+import { signInWithEmailAndPassword } from "firebase/auth";
 import { authService } from "../../utils/firebase";
 import { useState } from "react";
 import { useForm } from "react-hook-form";
+import SocialLogin from "./SocialLogin";
 
 interface LoginProps {
   closeModal: any;
@@ -50,32 +46,6 @@ const Login = ({ closeModal, setModal }: LoginProps) => {
         console.log("error: ", error);
         setIsLogin(false);
         setError("아이디와 비밀번호를 확인해주세요.");
-      });
-  };
-
-  // google login
-  const signInGoogle = () => {
-    const provider = new GoogleAuthProvider();
-    signInWithPopup(authService, provider)
-      .then((result) => {
-        alert("로그인 성공");
-        setModal(false);
-      })
-      .catch((error) => {
-        console.log("error: ", error);
-      });
-  };
-
-  // facebook login
-  const signInFacebook = () => {
-    const provider = new FacebookAuthProvider();
-    signInWithPopup(authService, provider)
-      .then((result) => {
-        alert("로그인 성공");
-        setModal(false);
-      })
-      .catch((error) => {
-        console.log("error: ", error);
       });
   };
 
@@ -168,30 +138,7 @@ const Login = ({ closeModal, setModal }: LoginProps) => {
           </div>
         </form>
       </div>
-      <div className="flex items-center justify-center p-6 border-t border-solid border-blueGray-200 rounded-b">
-        <div className="flex-column items-center">
-          <div>
-            <button onClick={signInGoogle}>
-              <img className="w-80" src="./login/google.png" alt="" />
-            </button>
-          </div>
-          <div>
-            <button onClick={signInFacebook}>
-              <img className="w-80" src="./login/facebook.png" alt="" />
-            </button>
-          </div>
-          {/* <div>
-            <button>
-              <img className="w-80" src="./login/kakao.png" alt="" />
-            </button>
-          </div>
-          <div>
-            <button>
-              <img className="w-80" src="./login/naver.png" alt="" />
-            </button>
-          </div> */}
-        </div>
-      </div>
+      <SocialLogin setModal={setModal} />
     </>
   );
 };
